+++ PAC/common/cip_tech_def.cpp
@@ -3476,7 +3476,8 @@ int cipline_tech_object::_CheckErr( )
     // Нет расхода на подаче
     if (T[TMR_NO_FLOW]->get_countdown_time() != (rt_par_float[P_TM_R_NO_FLOW] * 1000L))
         {
-        T[TMR_NO_FLOW]->set_countdown_time((unsigned long)rt_par_float[P_TM_R_NO_FLOW] * 1000L);
+        T[TMR_NO_FLOW]->set_countdown_time(
+            static_cast<uint32_t>( rt_par_float[P_TM_R_NO_FLOW] * 1000 ));
         }
     if (NP->get_state() == ON)
         {
@@ -3757,7 +3758,8 @@ int cipline_tech_object::InitFilRR( int where )
         divider = 1;
         }
     rt_par_float[P_MAX_OPER_TM] = (v/divider)*MAX_OP_TIME*3600;
-    T[TMR_OP_TIME]->set_countdown_time((unsigned long)rt_par_float[P_MAX_OPER_TM] * 1000L);
+    T[ TMR_OP_TIME ]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[ P_MAX_OPER_TM ] * 1000 ) );
     T[TMR_OP_TIME]->start();
     rt_par_float[P_SUM_OP] = 0;
     cnt->start();
@@ -3811,7 +3813,8 @@ int cipline_tech_object::InitCircRR( int where )
 
     rt_par_float[P_OP_TIME_LEFT] = 0;
     rt_par_float[P_MAX_OPER_TM] = parpar[0][P_TM_CIRC_RR];
-    T[TMR_OP_TIME]->set_countdown_time((unsigned long)rt_par_float[P_MAX_OPER_TM]*1000L);
+    T[ TMR_OP_TIME ]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[ P_MAX_OPER_TM ] * 1000 ) );
     T[TMR_OP_TIME]->start();
     rt_par_float[P_SUM_OP] = 0;
     cnt->start();
@@ -3823,7 +3826,8 @@ int cipline_tech_object::InitCheckConc( int where )
     InitCircRR(where);
     rt_par_float[P_MAX_OPER_TM] = parpar[0][P_TM_CHKC];
     T[TMR_OP_TIME]->reset();
-    T[TMR_OP_TIME]->set_countdown_time((unsigned long)rt_par_float[P_MAX_OPER_TM] * 1000L);
+    T[ TMR_OP_TIME ]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[ P_MAX_OPER_TM ] * 1000 ) );
     T[TMR_OP_TIME]->start();
     return 0;
     }
@@ -3919,7 +3923,8 @@ int cipline_tech_object::InitAddRR( int where, int step, int first_init_flag )
                 no_liquid_is_warning = 0;
                 break;
             }
-        T[TMR_OP_TIME]->set_countdown_time((unsigned long)rt_par_float[P_MAX_OPER_TM] * 1000L);
+        T[ TMR_OP_TIME ]->set_countdown_time(
+            static_cast<uint32_t>( rt_par_float[ P_MAX_OPER_TM ] * 1000 ) );
 
         rt_par_float[P_SUM_OP] = 0;
         }
@@ -3974,9 +3979,11 @@ int cipline_tech_object::InitOpolRR( int where )
 
     rt_par_float[P_OP_TIME_LEFT] = 0;
     rt_par_float[P_MAX_OPER_TM] = 300;
-    T[TMR_OP_TIME]->set_countdown_time((unsigned long)rt_par_float[P_MAX_OPER_TM] * 1000L);
+    T[ TMR_OP_TIME ]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[ P_MAX_OPER_TM ] * 1000 ) );
     T[TMR_OP_TIME]->start();
-    T[TMR_CHK_CONC]->set_countdown_time((unsigned long)rt_par_float[P_TM_NO_CONC] * 1000L);
+    T[TMR_CHK_CONC]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[P_TM_NO_CONC] * 1000 ) );
     rt_par_float[P_SUM_OP] = 0;
 
     rt_par_float[P_ZAD_CONC] = z;
@@ -4391,7 +4398,8 @@ int cipline_tech_object::_InitToObject( int from, int where, int step_to_init, i
         divider = 1;
         }
     rt_par_float[P_MAX_OPER_TM] = 3.6f * MAX_OP_TIME * p / divider;
-    T[TMR_OP_TIME]->set_countdown_time((unsigned long)rt_par_float[P_MAX_OPER_TM] * 1000);
+    T[ TMR_OP_TIME ]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[ P_MAX_OPER_TM ] * 1000 ) );
     T[TMR_OP_TIME]->start();
 
     rt_par_float[P_ZAD_CONC] = z;
@@ -4550,9 +4558,11 @@ int cipline_tech_object::_InitFromObject( int what, int where, int step_to_init,
         divider = 1;
         }
     rt_par_float[P_MAX_OPER_TM] = 3.6f * MAX_OP_TIME * p / divider;
-    T[TMR_OP_TIME]->set_countdown_time((unsigned long)rt_par_float[P_MAX_OPER_TM] * 1000);
+    T[ TMR_OP_TIME ]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[ P_MAX_OPER_TM ] * 1000 ) );
     T[TMR_OP_TIME]->start();
-    T[TMR_RETURN]->set_countdown_time((unsigned long)rt_par_float[P_TM_NO_FLOW_R] * 1000);
+    T[TMR_RETURN]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[P_TM_NO_FLOW_R] * 1000 ) );
     T[TMR_RETURN]->reset();
     T[TMR_CHK_CONC]->reset();
     if (what == TANK_K || what == TANK_S || where == TANK_S || where == TANK_K)
@@ -4637,9 +4647,11 @@ int cipline_tech_object::_InitOporCIP( int where, int step_to_init, int not_firs
     rt_par_float[P_VRAB] = 0;
     rt_par_float[P_OP_TIME_LEFT] = 0;
     rt_par_float[P_MAX_OPER_TM] = rt_par_float[P_TM_MAX_TIME_OPORCIP];
-    T[TMR_OP_TIME]->set_countdown_time((unsigned long)rt_par_float[P_MAX_OPER_TM]*1000);
+    T[ TMR_OP_TIME ]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[ P_MAX_OPER_TM ] * 1000 ) );
     T[TMR_OP_TIME]->start();
-    T[TMR_RETURN]->set_countdown_time((unsigned long)rt_par_float[P_TM_RET_IS_EMPTY]*1000);
+    T[TMR_RETURN]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[ P_TM_RET_IS_EMPTY ] * 1000 ) );
     T[TMR_RETURN]->reset();
     rt_par_float[P_SUM_OP] = 0;
 
@@ -4704,7 +4716,8 @@ int cipline_tech_object::_InitFilCirc( int with_what, int step_to_init, int f )
     rt_par_float[P_VRAB] = rt_par_float[P_V_RAB_ML];
     rt_par_float[P_OP_TIME_LEFT] = 0;
     rt_par_float[P_MAX_OPER_TM] = 300;
-    T[TMR_OP_TIME]->set_countdown_time((unsigned long)rt_par_float[P_MAX_OPER_TM]*1000);
+    T[ TMR_OP_TIME ]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[ P_MAX_OPER_TM ] * 1000 ) );
     T[TMR_OP_TIME]->start();
     return 0;
     }
@@ -4783,7 +4796,8 @@ int cipline_tech_object::_InitOporCirc( int where, int step_to_init, int not_fir
     rt_par_float[P_VRAB] = rt_par_float[P_V_LL_BOT];
     rt_par_float[P_OP_TIME_LEFT] = 0;
     rt_par_float[P_MAX_OPER_TM] = rt_par_float[P_TM_MAX_TIME_OPORBACHOK];
-    T[TMR_OP_TIME]->set_countdown_time((unsigned long)rt_par_float[P_MAX_OPER_TM]*1000);
+    T[ TMR_OP_TIME ]->set_countdown_time(
+        static_cast<uint32_t>( rt_par_float[ P_MAX_OPER_TM ] * 1000 ) );
     T[TMR_OP_TIME]->start();
     T[TMR_OP_TIME]->pause();
     return 0;
